Have offlineimap constantly running in the background
Mutt should be invoked with `mutto` (stands for mutt-offlineimap).  This starts
offlineimap in the background before mutt, and then stops it when mutt exits.
The advantage of having offlineimap run in the background is that it then only
requires SIGUSR1 to sync, meaning that the password does not need to be passed
to it each time.